Create component to render reservations for one car
